+++ scripts/macroGenerator.js
@@ -25,7 +25,14 @@ export async function generateMacro(
   let command = '';
 
   // Dependencies get checked first
-  command += genMacroDependencies(target, random, addSubtract, toggleRandom, toggleAddSubtract);
+  command += genMacroDependencies(
+    target,
+    random,
+    addSubtract,
+    toggleRandom,
+    toggleAddSubtract,
+    fields
+  );
 
   const str = (obj) => {
     if (!obj) return null;
@@ -39,7 +46,7 @@ const update = ${str(fields)};
   if (method === 'toggle') command += `const update2 = ${str(toggleFields)};\n\n`;
 
   // Insert Mass Edit control objects
-  if (random || addSubtract || toggleRandom || toggleAddSubtract) {
+  if (random || addSubtract || toggleRandom || toggleAddSubtract || hasSpecialField(fields)) {
     command += `\n// Mass Edit control objects\n`;
     command += `const randomizeFields = ${str(random)};\n`;
     command += `const addSubtractFields = ${str(addSubtract)};\n`;
@@ -168,16 +175,8 @@ const toggleOn = function (obj) {
 // ===============
 `;
 
-  // If toggling and macro running is enabled we need to insert trackers
-  if (runMacro || runMacroToggle) {
-    command += `
-const toggledOnPlaceables = [];
-const toggledOffPlaceables = [];
-`;
-  }
-
   // Update related code
-  if (random || addSubtract || toggleRandom || toggleAddSubtract) {
+  if (random || addSubtract || toggleRandom || toggleAddSubtract || hasSpecialField(fields)) {
     command += genUpdateWithMassEditDep(target, method, docName, runMacro || runMacroToggle);
   } else {
     command += genUpdate(target, method, docName, runMacro || runMacroToggle);
@@ -197,17 +196,20 @@ const advancedMacro = game.modules.get('advanced-macros')?.active;
 
   // Run macros if applicable
   if (runMacro) {
+    if (target === 'tokenGeneric' && method !== 'toggle') command += `const targets = [token];\n`;
+
+    const targets = method === 'toggle' ? 'toggleOnTargets' : 'targets';
     command += `
 // Apply macro
 const applyMacro = game.collections.get('Macro').find(m => m.name === '${runMacro}')
-if (applyMacro && toggledOnPlaceables.length) {
+if (applyMacro && ${targets}.length) {
   if(selectMacroTargets) {
     layer.activate();
     layer.releaseAll();
-    toggledOnPlaceables.forEach(t => t.control({ releaseOthers: false }));
+    ${targets}.forEach(t => t.control({ releaseOthers: false }));
   }
   if (advancedMacro) {
-    applyMacro.execute(toggledOnPlaceables);
+    applyMacro.execute(${targets});
   } else {
     applyMacro.execute({token, actor});
   }
@@ -221,14 +223,14 @@ if (applyMacro && toggledOnPlaceables.length) {
     command += `
 // Apply macro on toggle off
 const offMacro = game.collections.get('Macro').find(m => m.name === '${runMacroToggle}')
-if (offMacro && toggledOffPlaceables.length) {
+if (offMacro && toggleOffTargets.length) {
   if(selectToggleMacroTargets) {
     layer.activate();
     layer.releaseAll();
-    toggledOffPlaceables.forEach(t => t.control({ releaseOthers: false }));
+    toggleOffTargets.forEach(t => t.control({ releaseOthers: false }));
   }
   if (advancedMacro) {
-    offMacro.execute(toggledOffPlaceables);
+    offMacro.execute(toggleOffTargets);
   } else {
     offMacro.execute({token, actor});
   }
@@ -251,7 +253,14 @@ if (offMacro && toggledOffPlaceables.length) {
   }
 }
 
-function genMacroDependencies(target, random, addSubtract, toggleRandom, toggleAddSubtract) {
+function genMacroDependencies(
+  target,
+  random,
+  addSubtract,
+  toggleRandom,
+  toggleAddSubtract,
+  fields
+) {
   let dep = '';
 
   const depWarning = (module) => {
@@ -269,10 +278,9 @@ if (!game.modules.get('tagger')?.active) {
 
 `;
 
-  if (random || addSubtract || toggleRandom || toggleAddSubtract)
+  if (random || addSubtract || toggleRandom || toggleAddSubtract || hasSpecialField(fields))
     dep += `
 const MassEdit = game.modules.get('multi-token-edit');
-const GeneralDataAdapter = MassEdit?.api?.GeneralDataAdapter;
 if(!MassEdit?.active){
   ${depWarning('Mass Edit')}
   return;
@@ -286,18 +294,25 @@ if(!MassEdit?.active){
 function genUpdate(target, method, docName, macroTracking = false) {
   // Update related code
   let command = '';
+
+  if (method === 'toggle') {
+    command += `
+const toggleOnTargets = [];
+const toggleOffTargets = [];
+`;
+  }
+
   if (target === 'tokenGeneric') {
     if (method === 'update') {
       command += 'token.document.update(update);\n';
-      if (macroTracking) command += 'toggledOnPlaceables.push(token);\n';
     } else if (method === 'toggle') {
       command += `
 if (toggleOn(token, update)) {
   token.document.update(update2);
-  ${macroTracking ? 'toggledOffPlaceables.push(token);' : ''}
+  ${macroTracking ? 'toggleOffTargets.push(token);' : ''}
 } else {
   token.document.update(update);
-  ${macroTracking ? 'toggledOnPlaceables.push(token);' : ''}
+  ${macroTracking ? 'toggleOnTargets.push(token);' : ''}
 }
 `;
     }
@@ -309,10 +324,10 @@ targets.forEach((t) => {
   let u;
   if(toggleOn(t, update)) {
     u = deepClone(update2);
-    ${macroTracking ? 'toggledOffPlaceables.push(t);' : ''}
+    ${macroTracking ? 'toggleOffTargets.push(t);' : ''}
   } else {
     u = deepClone(update);
-    ${macroTracking ? 'toggledOnPlaceables.push(t);' : ''}
+    ${macroTracking ? 'toggleOnTargets.push(t);' : ''}
   }
   u._id = t.id;
   updates.push(u);
@@ -324,7 +339,6 @@ targets.forEach((t) => {
   let u = deepClone(update);
   u._id = t.id;
   updates.push(u);
-  ${macroTracking ? 'toggledOnPlaceables.push(t);' : ''}
 });
 `;
     }
@@ -349,128 +363,45 @@ function genUpdateWithMassEditDep(target, method, docName, macroTracking = false
 
   if (target === 'tokenGeneric') {
     if (method === 'update') {
-      command += `
-if (randomizeFields) MassEdit.api.applyRandomization([update], [token], randomizeFields);
-if (addSubtractFields) MassEdit.api.applyAddSubtract([update], [token], '${docName}', addSubtractFields);
-MassEdit.api.GeneralDataAdapter.formToData('${docName}', token, update);
-
-token.document.update(update);
-`;
-      if (macroTracking) command += 'toggledOnPlaceables.push(token);\n';
+      command += `await MassEdit.api.performMassUpdate.call({randomizeFields, addSubtractFields}, update, [token], '${docName}');`;
     } else if (method === 'toggle') {
       command += `
 if (toggleOn(token, update)) {
-  if (randomizeFieldsToggleOff) MassEdit.api.applyRandomization([update2], [token], randomizeFieldsToggleOff);
-  if (addSubtractFieldsToggleOff) MassEdit.api.applyAddSubtract([update2], [token], '${docName}', addSubtractFieldsToggleOff);
-  MassEdit.api.GeneralDataAdapter.formToData('${docName}', token, update2);
-
-  token.document.update(update2);
-  ${macroTracking ? 'toggledOffPlaceables.push(token);' : ''}
+  await MassEdit.api.performMassUpdate.call({randomizeFields, addSubtractFields}, update2, [token], '${docName}');
 } else {
-  if (randomizeFields) MassEdit.api.applyRandomization([update], [token], randomizeFields);
-  if (addSubtractFields) MassEdit.api.applyAddSubtract([update], [token], '${docName}', addSubtractFields);
-  MassEdit.api.GeneralDataAdapter.formToData('${docName}', token, update);
-
-  token.document.update(update);
-  ${macroTracking ? 'toggledOnPlaceables.push(token);' : ''}
+  await MassEdit.api.performMassUpdate.call({randomizeFields, addSubtractFields}, update, [token], '${docName}');
 }
 `;
     }
   } else {
-    command += '\nconst updates = [];';
-
     if (method === 'toggle') {
       command += `
 const toggleOnTargets = [];
-const toggleOffUpdates = [];
 const toggleOffTargets = [];
 
 targets.forEach((t) => {
-  let u;
-  if(toggleOn(t, update)) {u = deepClone(update2);
-    toggleOffUpdates.push(u);
+  if(toggleOn(t, update)) {
     toggleOffTargets.push(t);
-    ${macroTracking ? 'toggledOffPlaceables.push(t);' : ''}
   } else {
-    u = deepClone(update);
     toggleOnTargets.push(t);
-    ${macroTracking ? 'toggledOnPlaceables.push(t);' : ''}
-    updates.push(u)
   }
-  u._id = t.id;
 });
-`;
-    } else {
-      command += `
-targets.forEach((t) => {
-  let u = deepClone(update);
-  u._id = t.id;
-  updates.push(u);
-  ${macroTracking ? 'toggledOnPlaceables.push(t);' : ''}
-});
-`;
-    }
-
-    const toggleOn = method === 'toggle' ? 'toggleOnTargets' : 'targets';
-
-    command += `
-if (updates.length) {
-  if (randomizeFields) MassEdit.api.applyRandomization(updates, ${toggleOn}, randomizeFields);
-  if (addSubtractFields) MassEdit.api.applyAddSubtract(updates, ${toggleOn}, '${docName}', addSubtractFields);
-}
-`;
-
-    if (method === 'toggle') {
-      command += `
-if (toggleOffUpdates.length) {
-  if (randomizeFieldsToggleOff) MassEdit.api.applyRandomization(toggleOffUpdates, toggleOffTargets, randomizeFieldsToggleOff);
-  if (addSubtractFieldsToggleOff) massEdit.api.applyAddSubtract(toggleOffUpdates, toggleOffTargets, '${docName}', addSubtractFieldsToggleOff);
-}
 `;
     }
 
     if (method === 'toggle') {
       command += `
-for (let i = 0; i < updates.length; i++) {
-  GeneralDataAdapter.formToData('${docName}', toggleOnTargets[i], updates[i]);
-  GeneralDataAdapter.formToData('${docName}', toggleOffTargets[i], toggleOffUpdates[i]);
-}
+await MassEdit.api.performMassUpdate.call({randomizeFields, addSubtractFields}, update, toggleOnTargets, '${docName}');
+await MassEdit.api.performMassUpdate.call({randomizeFields, addSubtractFields}, update2, toggleOffTargets, '${docName}');
 `;
-      if (SUPPORTED_PLACEABLES.includes(docName)) {
-        command += `\ncanvas.scene.updateEmbeddedDocuments('${docName}', updates.concat(toggleOffUpdates));`;
-      } else if (docName === 'PlaylistSound') {
-        command += `
-const allUpdates = updates.concat(toggleOffUpdates);
-for (let i = 0; i < allUpdates.length; i++) {
-  const t = targets.find(t => t.id === allUpdates[i]._id);
-  delete allUpdates[i]._id;
-  t.document.update(allUpdates[i]);
-}
-`;
-      } else {
-        command += `\ntargets[0]?.document.constructor?.updateDocuments(updates.concat(toggleOffUpdates))`;
-      }
     } else {
-      command += `
-for (let i = 0; i < updates.length; i++) {
-  GeneralDataAdapter.formToData('${docName}', targets[i], updates[i]);
-}
-`;
-      if (SUPPORTED_PLACEABLES.includes(docName)) {
-        command += `\ncanvas.scene.updateEmbeddedDocuments('${docName}', updates)`;
-      } else if (docName === 'PlaylistSound') {
-        command += `
-for (let i = 0; i < updates.length; i++) {
-  const t = targets.find(t => t.id === updates[i]._id);
-  delete updates[i]._id;
-  t.document.update(updates[i]);
-}
-`;
-      } else {
-        command += `\ntargets[0]?.document.constructor?.updateDocuments(updates)`;
-      }
+      command += `await MassEdit.api.performMassUpdate.call({randomizeFields, addSubtractFields}, update, targets, '${docName}');`;
     }
   }
 
   return command;
+}
+
+export function hasSpecialField(fields) {
+  return 'tokenmagic.ddTint' in fields || 'tokenmagic.preset' in fields;
 }